+++ pkg/middleware/recovery_test.go
@@ -36,7 +36,7 @@ func TestRecoveryMiddleware(t *testing.T) {
 
 			So(sc.resp.Code, ShouldEqual, 500)
 			So(sc.resp.Header().Get("content-type"), ShouldEqual, "text/html; charset=UTF-8")
-			So(sc.resp.Body.String(), ShouldContainSubstring, "<title>Grafana - Error</title>")
+			So(sc.resp.Body.String(), ShouldContainSubstring, "<title>Synectiks - Error</title>")
 		})
 	})
 }